[#884] - Newly created record disappear during syncing

+++ platform/shared/rubyJVM/src/com/rho/sync/SyncSource.java
@@ -99,6 +99,7 @@ class SyncSource
     int  getTotalCount(){return m_nTotalCount;}
     
     SyncEngine getSync(){ return m_syncEngine; }
+    SyncNotify getNotify(){ return getSync().getNotify(); }
 	DBAdapter getDB(){ return getSync().getDB(); }
 	NetRequest getNet(){ return getSync().getNet(); }
 
@@ -132,11 +133,14 @@ class SyncSource
     
 	void sync() throws Exception
 	{
-    	m_syncEngine.fireNotification(this, false, RhoRuby.ERR_NONE, "Synchronizing " + getName() + "...");
+    	getNotify().fireSyncNotification(this, false, RhoRuby.ERR_NONE, "Synchronizing " + getName() + "...");
 		
 	    TimeInterval startTime = TimeInterval.getCurrentTime();
 	    
 	    try{
+	        if ( isEmptyToken() )
+	            processToken("1");
+	    	
 	        if ( m_strParams.length() == 0 )
 	        {
 			    syncClientChanges();
@@ -224,7 +228,7 @@ class SyncSource
 	        {
 			    LOG.INFO( "Push blobs to server. Source id: " + getID() + "Count :" + m_arSyncBlobs.size() );
 
-	            if ( i <= 1 ) //create, update
+	            if ( getSync().SYNC_VERSION() < 2 && i <= 1 ) //create, update
 	            {
 	                IDBResult res = getDB().executeSQL("SELECT object, attrib " +
 						     "FROM changed_values WHERE source_id=? and update_type=? and (attrib_type IS NULL or attrib_type!=?)", 
@@ -243,12 +247,17 @@ class SyncSource
 	                }
 	            }
 
-	            getDB().executeSQL("DELETE FROM changed_values WHERE source_id=? and update_type=? and (attrib_type IS NULL or attrib_type!=?)", 
-	                getID(), arUpdateTypes[i], "blob.file" );
+	            if ( getSync().SYNC_VERSION() >= 2 && i == 1 )// update
+	                getDB().executeSQL("UPDATE changed_values SET update_type='update_sent' WHERE source_id=? and update_type=? and (attrib_type IS NULL or attrib_type!=?)", 
+	                    getID(), arUpdateTypes[i], "blob.file" );
+	            else
+	            	getDB().executeSQL("DELETE FROM changed_values WHERE source_id=? and update_type=? and (attrib_type IS NULL or attrib_type!=?)", 
+	            		getID(), arUpdateTypes[i], "blob.file" );
+	            
 	            syncClientBlobs(strUrl+strQuery);
 	        }else if ( strBody.length() > 0 )
 	        {
-	        	if ( i <= 1 ) //create, update
+	        	if ( getSync().SYNC_VERSION() < 2 && i <= 1 ) //create, update
 	            {
 	                IDBResult res = getDB().executeSQL("SELECT object, attrib "+
 						     "FROM changed_values where source_id=? and update_type =?", getID(), arUpdateTypes[i] );
@@ -264,8 +273,10 @@ class SyncSource
 	                    getDB().endTransaction();
 	                }
 	            }
-
-	            getDB().executeSQL("DELETE FROM changed_values WHERE source_id=? and update_type=?", getID(), arUpdateTypes[i] );
+	            if ( getSync().SYNC_VERSION() >= 2 && i == 1 )// update
+	                getDB().executeSQL("UPDATE changed_values SET update_type='update_sent' WHERE source_id=? and update_type=?", getID(), arUpdateTypes[i] );
+	            else
+	            	getDB().executeSQL("DELETE FROM changed_values WHERE source_id=? and update_type=?", getID(), arUpdateTypes[i] );
 	        }
 	    }
 	}
@@ -352,7 +363,7 @@ class SyncSource
 	            strUrl += '/' + m_strAction;
 	        
 			String strQuery = SyncEngine.SYNC_SOURCE_FORMAT() + "&client_id=" + getSync().getClientID() + 
-	                "&p_size=" + SyncEngine.SYNC_PAGE_SIZE() + "&version=1";
+	                "&p_size=" + SyncEngine.SYNC_PAGE_SIZE() + "&version=" + getSync().SYNC_VERSION();
 	        if ( m_strParams.length() > 0 )
 	            strQuery += m_strParams;
 	
@@ -363,9 +374,9 @@ class SyncSource
 	            strQuery += "&question=" + getAskParams();
 	        }
 	
-	        if ( isEmptyToken() )
+	        /*if ( isEmptyToken() )
 	            processToken("1");
-	        else if ( !m_bTokenFromDB && isTokenMoreThanOne() )
+	        else*/ if ( !m_bTokenFromDB && isTokenMoreThanOne() )
 	            strQuery += "&ack_token=" + getToken();
 	
 			LOG.INFO( "Pull changes from server. Url: " + (strUrl+strQuery) );
@@ -392,17 +403,28 @@ class SyncSource
 	
 	        processServerData(resp.getCharData());
 	        
-	        /*String strData =
-			"[{count: 124},{version: 1},{total_count: 5425},{token: 123}," +
-	        "{s:\"AeropriseRequest\",ol:[" +
-			"{o:\"2ed2e0c7-8c4c-99c6-1b37-498d250bb8e7\",av:[" +
-			"{a:\"first_name\",i:47354289,v:\"Lars. \n\n Burgess\", t:\"blob\"}," +
-	        "{a:\"second_name\",i:55555,v:\"Burgess\"}" +
-	        "]}" +
+			//String strData =
+	        //"[{count:10},{version:1},{total_count: 5425},{token: 123},{s:\"RhoDeleteSource\",ol:[{o:\"rho_del_obj\",av:[{i:55550425},{i:75665819},{i:338165272},{i:402396629},{i:521753981},{i:664143530},{i:678116186},{i:831092394},{i:956041217},{i:970452458}]}]}]";
+			/*"[{count: 124},{version: 1},{total_count: 5425},{token: 123},"
+	        "{s:\"Product\",ol:["
+			"{oo:\"123\",o:\"2ed2e0c7-8c4c-99c6-1b37-498d250bb8e7\",av:["
+			"{a:\"first_name\",i:47354289,v:\"Lars. \n\n Burgess\", t:\"blob\"},"
+	        "{a:\"second_name\",i:55555,v:\"Burgess\"}]},"
+	        "{oo:\"456\", e:\"Something went wrong creating this record on the backend: code 7\"}"
+	        "]}]"; */
+			/*"[{count: 1},{version: 1},{total_count: 1},{token: 123},"
+	        "{s:\"Product\",ol:["
+	        "{oo:\"94\", e:\"Something went wrong creating this record on the backend: code 7\"}"
+	        "]}]";*/
+			/*"[{count: 1},{version: 1},{total_count: 1},{token: 123},"
+	        "{s:\"Product\",ol:["
+	        "{o:\"94\", av:["
+	        "{a:\"TEST\",i:55555,v:\"Geny\"}]},"
 	        "]}]";
-			//u:\"query\",
-	        processServerData(strData);*/
-	        
+
+			//u:\"query\",  
+			processServerData(strData.c_str()); */
+
 	        if ( getAskParams().length() > 0 || getCurPageCount() == 0 )
 	            break;
 	    }
@@ -418,7 +440,7 @@ class SyncSource
 	    {
 	        m_strError = oJsonArr.getCurItem().getString("error");
 	        m_nErrCode = RhoRuby.ERR_CUSTOMSYNCSERVER;
-	        processToken("0");
+	        //processToken("0");
 	        getSync().stopSync();
 	        return;
 	    }
@@ -442,7 +464,7 @@ class SyncSource
 	    }
 	    
 	    if ( getServerObjectsCount() == 0 )
-	    	m_syncEngine.fireNotification(this, false, RhoRuby.ERR_NONE, "");
+	    	getNotify().fireSyncNotification(this, false, RhoRuby.ERR_NONE, "");
 	    
 	    if ( !oJsonArr.isEnd() )
 	    {
@@ -470,13 +492,15 @@ class SyncSource
 		    	
 		    	PROF.START("DB");
 				getDB().endTransaction();
-			    PROF.STOP("DB");		    
+			    PROF.STOP("DB");
+			    
+			    getNotify().fireObjectsNotification();
 			}
 		    
 		}
 		
-	    if ( getServerObjectsCount() < getTotalCount() )
-	    	m_syncEngine.fireNotification(this, false, RhoRuby.ERR_NONE, "");
+		if ( getCurPageCount() > 0 )
+	    	getNotify().fireSyncNotification(this, false, RhoRuby.ERR_NONE, "");
 	}
 
 	void processServerData_Ver0(JSONArrayIterator oJsonArr)throws Exception
@@ -516,7 +540,8 @@ class SyncSource
 	            if ( pSrc == null )
 	            {
 	                LOG.ERROR("Sync server send data for unknown source name:" + strSrcName);
-	                continue;
+	                getSync().stopSync();
+	                break;
 	            }
 	            nSrcID = pSrc.getID();
 	        }
@@ -547,13 +572,13 @@ class SyncSource
 	{
 	    String m_strValue;
 		String m_strAttrType;
-		long   m_nID;
+		Long   m_nID;
 		
 		CValue(JSONEntry oJsonEntry)throws JSONException
 		{
 			m_strValue = oJsonEntry.getString("value");
 			m_strAttrType = oJsonEntry.getString("attrib_type");
-			m_nID = oJsonEntry.getLong("id");
+			m_nID = new Long(oJsonEntry.getLong("id"));
 		}
 		CValue(JSONEntry oJsonEntry, int nVer)throws JSONException
 		{
@@ -561,7 +586,7 @@ class SyncSource
 			{
 				m_strValue = oJsonEntry.getString("v");
 				m_strAttrType = oJsonEntry.getString("t");
-				m_nID = oJsonEntry.getLong("i");
+				m_nID = new Long(oJsonEntry.getLong("i"));
 			}
 		}
 		
@@ -587,7 +612,7 @@ class SyncSource
 			}
 		}
 		
-		fName += "/id_" + Long.toString(value.m_nID) + strExt;
+		fName += "/id_" + value.m_nID.toString() + strExt;
 		
 		return fName;
 	}
@@ -631,7 +656,16 @@ class SyncSource
 	
 	boolean processSyncObject_ver1(JSONEntry oJsonObject, Integer nSrcID)throws Exception
 	{
+	    if ( oJsonObject.hasName("e") )
+	    {
+	    	String strOldObject = oJsonObject.getString("oo");
+	    	String strError = oJsonObject.getString("e");
+	        getNotify().addCreateObjectError(nSrcID,strOldObject,strError);
+	        return true;
+	    }
+		
 		String strObject = oJsonObject.getString("o");
+		String strOldObject = oJsonObject.getString("oo");
 		JSONArrayIterator oJsonArr = new JSONArrayIterator(oJsonObject, "av");
 		
 	    for( ; !oJsonArr.isEnd() && getSync().isContinueSync(); oJsonArr.next() )
@@ -652,21 +686,57 @@ class SyncSource
 		    	//if( oJsonEntry.hasName("u") )
 		    	//	strUpdateType = oJsonEntry.getString("u");
 		    	
-		        getDB().executeSQL("INSERT INTO object_values " +
-		            "(id, attrib, source_id, object, value, attrib_type) VALUES(?,?,?,?,?,?)", 
-		            new Long(value.m_nID), strAttrib, nSrcID, strObject,
-		            value.m_strValue, value.m_strAttrType );
-		        
+	            boolean bUpdated = false;
+	            if ( strOldObject != null )
+	            {
+	                IDBResult res = getDB().executeSQL("SELECT object FROM object_values where object=? and attrib=? and source_id=?", strOldObject, strAttrib, nSrcID );
+	                if ( !res.isEnd() )
+	                {
+	                    getDB().executeSQL("UPDATE object_values SET id=?, object=? where object=? and attrib=? and source_id=?", value.m_nID, strObject, strOldObject, strAttrib, nSrcID );
+	                    getNotify().onObjectChanged(nSrcID,strOldObject, SyncNotify.enCreate);
+
+	                    bUpdated = true;
+	                }
+	            }
+
+	            if ( !bUpdated )
+	            {
+	                IDBResult res = getDB().executeSQL("SELECT value, attrib_type " +
+						     "FROM changed_values where object=? and attrib=? and source_id=? LIMIT 1 OFFSET 0", strObject, strAttrib, nSrcID );
+	                if ( !res.isEnd() )
+	                {
+	                    boolean bModified = false;
+	                    if ( res.getStringByIdx(0).compareTo(value.m_strValue) != 0 || res.getStringByIdx(1).compareTo(value.m_strAttrType) != 0 )
+	                        bModified = true;
+
+	                    getDB().executeSQL("UPDATE object_values SET id=?, value=?, attrib_type=? where object=? and attrib=? and source_id=?", 
+	                        value.m_nID, value.m_strValue, value.m_strAttrType, strObject, strAttrib, nSrcID );
+	                    getDB().executeSQL("DELETE FROM changed_values where object=? and attrib=? and source_id=?", strObject, strAttrib, nSrcID );
+
+	                    if ( bModified )
+	                        getNotify().onObjectChanged(nSrcID,strObject, SyncNotify.enUpdate);
+	                }
+	                else
+	    		        getDB().executeSQL("INSERT INTO object_values " +
+	    			            "(id, attrib, source_id, object, value, attrib_type) VALUES(?,?,?,?,?,?)", 
+	    			            value.m_nID, strAttrib, nSrcID, strObject,
+	    			            value.m_strValue, value.m_strAttrType );
+	            }
+		    	
 		        m_nInserted++;
 		    }else// if ( nDbOp == 1 ) //delete
 		    {
 		    	long id = oJsonEntry.getLong("i");
-	            IDBResult res = getDB().executeSQL("SELECT source_id FROM object_values where id=?", id );
+		    	
+	            IDBResult res = getDB().executeSQL("SELECT source_id, object FROM object_values where id=?", id );
 	            if ( !res.isEnd() )
 	            {
-			        getDB().executeSQL("DELETE FROM object_values where id=?", id );
+	                Integer nDelSrcID = new Integer(res.getIntByIdx(0));
+	                String strDelObject = res.getStringByIdx(1);
+	                getDB().executeSQL("DELETE FROM object_values where id=?", id );
+	                getNotify().onObjectChanged(nDelSrcID, strDelObject, SyncNotify.enDelete);
 	            }
-	            
+		    	
 		        m_nDeleted++;
 		    }//else{
 		     //   LOG.ERROR("Unknown DB operation: " + nDbOp );
@@ -697,7 +767,7 @@ class SyncSource
 	    	String strAttrib = oJsonEntry.getString("attrib");
 	        getDB().executeSQL("INSERT INTO object_values " +
 	            "(id, attrib, source_id, object, value, attrib_type) VALUES(?,?,?,?,?,?)", 
-	            new Long(value.m_nID), strAttrib, getID(), oJsonEntry.getString("object"),
+	            value.m_nID, strAttrib, getID(), oJsonEntry.getString("object"),
 	            value.m_strValue, value.m_strAttrType );
 	        
 	        m_nInserted++;
@@ -730,112 +800,3 @@ class SyncSource
 	
 	}
 }
-
-/*
-void syncClientChanges()throws Exception
-{
-	LOG.INFO("Sync client changes source ID :" + getID() );
-	
-    IDBResult res = getDB().executeSQL("SELECT attrib, object, value, attrib_type, update_type "+
-			 "FROM object_values where source_id=? and (update_type = 'update' or update_type = 'create' or update_type = 'delete') order by update_type", getID() );
-    
-    String strUpdateType = "";
-	m_arSyncBlobs.removeAllElements();
-	m_strPushBody = "";
-	
-    for( ; !res.isEnd()&& getSync().isContinueSync(); res.next() )
-    {
-    	String strTemp  = res.getStringByIdx(4);
-    	if ( !strUpdateType.equals(strTemp) )
-    	{
-    		if ( strUpdateType.length() > 0 )
-    		{
-    			if ( !sendClientChanges(strUpdateType) )
-    			{
-    				getSync().setState(SyncEngine.esStop);
-                	continue;
-    			}	
-    		}
-    		
-    		m_strPushBody = "";
-    		strUpdateType = strTemp;
-    	}
-    	
-    	makePushBody( res );
-    }
-    
-    if ( getSync().isContinueSync() && strUpdateType.length() > 0 )
-    	sendClientChanges(strUpdateType);
-}
-
-boolean sendClientChanges(String strUpdateType)throws Exception
-{
-    String strUrl = getUrl() + "/" + strUpdateType;
-    strUrl += "objects";
-    String strQuery = SyncEngine.SYNC_SOURCE_FORMAT() + "&client_id=" + getSync().getClientID();
-
-    if ( m_strPushBody.length() > 0 )
-    {
-	    LOG.INFO("Push client changes to server. Source id: " + getID() + "Size :" + m_strPushBody.length());
-	    LOG.TRACE( "Push body: " + m_strPushBody );		
-
-	    try{
-	    	NetResponse resp = getNet().pushData(strUrl+strQuery,m_strPushBody, getSync() );
-	        if ( !resp.isOK() )
-	        {
-	        	m_nErrCode = RhoRuby.ERR_REMOTESERVER;
-	        	//m_strError = resp.getCharData();
-	        	
-	            return false;
-	        }
-	    }catch(Exception exc)
-	    {
-	    	m_nErrCode = RhoRuby.getNetErrorCode(exc);
-	    	throw exc;
-	    }
-    }
-
-    if ( m_arSyncBlobs.size() > 0  )
-    {
-	    LOG.INFO( "Push blobs to server. Source id: " + getID() + "Count :" + m_arSyncBlobs.size() );
-
-        getDB().executeSQL("DELETE FROM object_values WHERE source_id=? and update_type=? and (attrib_type IS NULL or attrib_type!=?)", getID(), strUpdateType, "blob.file" );
-        syncClientBlobs(strUrl+strQuery);
-    }else if ( m_strPushBody.length() > 0 )
-        getDB().executeSQL("DELETE FROM object_values WHERE source_id=? and update_type=?", getID(), strUpdateType );
-    
-    return true;
-}
-*/	
-/*	void makePushBody( IDBResult res )throws DBException
-{
-    String strSrcBody = "attrvals[][attrib]=" + res.getStringByIdx(0);
-
-    if ( res.getStringByIdx(1).length() > 0 ) 
-        strSrcBody += "&attrvals[][object]=" + res.getStringByIdx(1);
-
-    String value = res.getStringByIdx(2);
-    String attribType = res.getStringByIdx(3);
-
-    //if ( value.length() > 0 )
-    {
-        if ( attribType.equals("blob.file") )
-        {
-            FilePath oBlobPath = new FilePath(value);
-            strSrcBody += "&attrvals[][value]=";
-            strSrcBody += oBlobPath.getBaseName();
-            strSrcBody += "&attrvals[][attrib_type]=blob";
-
-            if ( value.length() > 0 )
-            	m_arSyncBlobs.addElement(new SyncBlob(strSrcBody,value));
-            return;
-        }else
-            strSrcBody += "&attrvals[][value]=" + value;
-    }
-
-    if ( m_strPushBody.length() > 0 )
-    	m_strPushBody += "&";
-
-    m_strPushBody += strSrcBody;
-}
-*/
